- Change Healthbar setting Max and Min variable as private, and Max health set 50.

+++ VGP114_Super_Hot_Proj/Assets/Scripts/Enemy.cs
@@ -93,7 +93,7 @@ public class Enemy : MonoBehaviour, ICharacterAction
         Gizmos.DrawLine(position, position + 3.0f * transform.forward);
     }
 
-    public void TakeDamage(float dmg)
+    public void TakeDamage(int dmg)
     {
         health -= dmg;
         if (health <= 0.0f)
